Fix build issue of dynamic client updation patch on Android-P
While building the Configure client dynamically without restarting the
AIC server patch on Android-P, build issue was seen because of type
mismatch of cameraId iterator and return type of size of map.

While iterating through size of map using an int type iterator,
it is required to typecast the size of map to int type.

Tracked-On: OAM-104587

+++ src/VirtualCameraFactory.cpp
@@ -256,7 +256,7 @@ bool VirtualCameraFactory::createVirtualRemoteCamera(
     ALOGV("%s: E", __FUNCTION__);
     int cameraId = mVirtualCameras.size();
     std::shared_ptr<ClientCommunicator> client_thread = mClientThreads[clientId];
-    for (int id = 0; id < mVirtualCameras.size(); id++) {
+    for (int id = 0; id < (int)mVirtualCameras.size(); id++) {
         if (mVirtualCameras[id] == nullptr) {
             ALOGI("%s:CameraId is set to %d", __FUNCTION__, id);
             cameraId = id;
